Fix inference-loss bug
In the model forward function, we had

```
if self.training:
    <classification slice is left alone>
else:
    <sigmoid the classification slice>
```

when measuring validation / test error, we .eval() the model,
so we sigmoid the classification loss. For CrossEntropyError,
we do not want to do that. Therefore we got lotsa loss for
test/val that doesn't actually exist. The models should have
trained fine, though!